refactor: turn ScrollToTop into custom hook
