+++ .env
@@ -1,5 +1,5 @@
 REACT_APP_APIPATH=https://dimsserver.azurewebsites.net/api
-REACT_APP_FIREBASE_APIKEY=
-REACT_APP_FIREBASE_PROJECTID=
-REACT_APP_FIREBASE_SENDERID=
-REACT_APP_FIREBASE_APPID=
+REACT_APP_FIREBASE_APIKEY=AIzaSyDICQVta0ppDHQgm07f0Odeehkm2y4j0cA
+REACT_APP_FIREBASE_PROJECTID=dimsui-ae0a4
+REACT_APP_FIREBASE_SENDERID=521943275729
+REACT_APP_FIREBASE_APPID=1:521943275729:web:ebb2a955d4d72046430df0
