+++ src/components/elements/Form/Input/CustomInputs/TextArea.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import checkCharsCount from '../../../helpers/formHelpers/checkCharsCount';
-import compareObjects from '../../../helpers/compareObjects';
+import checkCharsCount from '../../../../../helpers/formHelpers/checkCharsCount';
+import compareObjects from '../../../../../helpers/compareObjects';
 
 function TextArea(props) {
   const {
