+++ src/helpers/hooks/useScrollToTop.js
@@ -1,0 +1,11 @@
+import { useEffect } from 'react';
+
+const useScrollToTop = ({ location }) => {
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, [location]);
+
+  return null;
+};
+
+export default useScrollToTop;
